ai_service/requirements.txt
Flask
scikit-learn
pandas
numpy
joblib
pymongo
python-dotenv